backend/requirements.txt
langgraph>=0.2.0
fastapi>=0.104.0
uvicorn>=0.24.0
google-generativeai>=0.3.0
langchain-google-genai>=1.0.0
langchain-community>=0.0.20
langchain>=0.1.0
langgraph>=0.0.40
pydantic>=2.5.0
python-dotenv>=1.0.0
requests>=2.31.0
httpx>=0.25.0